backend/requirements.txt
functions-framework==3.5.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-generativeai==0.3.2
reportlab==4.0.7
requests==2.31.0
flask==3.0.0